Move UrlUtils to a standalone module
