+++ app/build.gradle
@@ -196,6 +196,7 @@ dependencies {
     implementation project(':feature-tabs')
     implementation project(':HttpRequest')
     implementation project(':httptask')
+    implementation project(':urlutils')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
