+++ app/src/main/java/org/mozilla/focus/fragment/BrowserFragment.java
@@ -68,8 +68,8 @@ import org.mozilla.focus.utils.IntentUtils;
 import org.mozilla.focus.utils.Settings;
 import org.mozilla.focus.utils.SupportUtils;
 import org.mozilla.focus.utils.ThreadUtils;
-import org.mozilla.focus.utils.UrlUtils;
 import org.mozilla.focus.web.GeoPermissionCache;
+import org.mozilla.urlutils.UrlUtils;
 import org.mozilla.focus.widget.AnimatedProgressBar;
 import org.mozilla.focus.widget.BackKeyHandleable;
 import org.mozilla.focus.widget.FragmentListener;
