+++ app/src/webkit/java/org/mozilla/focus/webkit/WebViewDebugOverlay.java
@@ -28,7 +28,7 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import org.mozilla.focus.BuildConfig;
-import org.mozilla.focus.utils.UrlUtils;
+import org.mozilla.urlutils.UrlUtils;
 
 public class WebViewDebugOverlay {
     private WebView webView;
