+++ components/utils/urlutils/src/test/java/org/mozilla/urlutils/UrlUtilsTest.java
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.focus.utils;
+package org.mozilla.urlutils;
 
 import android.annotation.SuppressLint;
